+++ src/main/java/com/lovesoongalarm/lovesoongalarm/domain/websocket/sub/subscription/business/UserSubscriptionService.java
@@ -3,7 +3,6 @@ package com.lovesoongalarm.lovesoongalarm.domain.websocket.sub.subscription.busi
 import com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.sub.message.application.dto.UserChatUpdateDTO;
 import com.lovesoongalarm.lovesoongalarm.domain.websocket.sub.messaging.MessageSender;
 import com.lovesoongalarm.lovesoongalarm.domain.websocket.sub.session.SessionService;
-import com.lovesoongalarm.lovesoongalarm.domain.websocket.sub.subscription.implement.RedisSubscriber;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
